Initial version

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -2,15 +2,15 @@ package com.kodilla.com.kodilla.basic_assertion;
 
 public class Calculator {
 
-    public int sum(int a, int b) {
-        return a + b;
+    public int sum(double a, double b) {
+        return (int) (a + b);
     }
 
     public int subtract(int a, int b) {
         return a - b;
     }
 
-    public int exponentiation(int a, int n) {
+    public int exponentiation(double a, double n) {
         return (int) Math.pow(a, n);
     }
 }
